+++ src/utilities/base-path.ts
@@ -10,11 +10,11 @@ export function setBasePath(path: string) {
  *
  * The base path is used to load assets such as icons and images, so it needs to be set for components to work properly.
  * By default, this script will look for a script ending in webawesome.js or autoloader.js and set the base path to the
- * directory that contains that file. To override this behavior, you can add the data-webawesome attribute to any script
+ * directory that contains that file. To override this behavior, you can add the data-web-awesome attribute to any script
  * on the page (it probably makes the most sense to attach it to the Web Awesome script, but it could also be on a
  * bundle). The value can be a local folder or it can point to a CORS-enabled endpoint such as a CDN.
  *
- *   <script src="bundle.js" data-webawesome="/custom/base/path"></script>
+ *   <script src="bundle.js" data-web-awesome="/custom/base/path"></script>
  *
  * Alternatively, you can set the base path manually using the exported setBasePath() function.
  *
@@ -23,11 +23,11 @@ export function setBasePath(path: string) {
 export function getBasePath(subpath = '') {
   if (!basePath) {
     const scripts = [...document.getElementsByTagName('script')] as HTMLScriptElement[];
-    const configScript = scripts.find(script => script.hasAttribute('data-webawesome'));
+    const configScript = scripts.find(script => script.hasAttribute('data-web-awesome'));
 
     if (configScript) {
-      // Use the data-webawesome attribute
-      setBasePath(configScript.getAttribute('data-webawesome')!);
+      // Use the data-web-awesome attribute
+      setBasePath(configScript.getAttribute('data-web-awesome')!);
     } else {
       const fallbackScript = scripts.find(s => {
         return /webawesome(\.min)?\.js($|\?)/.test(s.src) || /autoloader(\.min)?\.js($|\?)/.test(s.src);